Make some integer casts explicit

+++ src/gui/qt/plugin_item_model.cpp
@@ -620,7 +620,7 @@ void PluginItemModel::handleHideMessage(const std::string& pluginName,
 
     for (size_t i = 0; i < items.size(); i += 1) {
       if (items.at(i).name == pluginName) {
-        auto index = this->index(i + 1, CARDS_COLUMN);
+        auto index = this->index(static_cast<int>(i) + 1, CARDS_COLUMN);
         emit dataChanged(index, index, {FilteredContentRole});
         break;
       }
